+++ tests/testthat/test-utils.R
@@ -41,7 +41,8 @@ test_that("check_columns", {
     ))
     df_result <- data.frame(
         ColN1 = c(1, 2), ColN2 = 4,
-        ColTU1 = "A", ColTU2 = 3, ColU1 = NA, ColU2 = NA
+        ColTU1 = "A", ColTU2 = 3,
+        ColU1 = NA_character_, ColU2 = NA_character_
     )
     expect_equal(df_get, df_result)
 
@@ -52,7 +53,8 @@ test_that("check_columns", {
     ))
     df_result <- data.frame(
         ColN1 = c(1, 2), ColN2 = 4,
-        ColTU1 = "A", ColTU2 = 3, ColTU3 = NA
+        ColTU1 = "A", ColTU2 = 3,
+        ColTU3 = NA_character_
     )
     expect_equal(df_get, df_result)
 })
@@ -199,3 +201,42 @@ test_that("plink_to_pedigree", {
     expect_equal(class(pedi), class(Pedigree()))
     expect_equal(length(pedi), 6)
 })
+
+test_that("complete_twins", {
+    data("relped")
+    df <- complete_twins(relped)
+    expect_equal(dim(df), c(9, 5))
+
+    rel_df <- data.frame(
+        id1 = c(112, 113, 133, 209),
+        id2 = c(110, 114, 132, 109),
+        code = c(1, 4, 4, 4)
+    )
+    df <- complete_twins(rel_df)
+    expect_equal(dim(df), c(4, 5))
+
+    rel_df <- data.frame(
+        id1 = c(112, 113, 133, 209),
+        id2 = c(110, 114, 114, 109),
+        code = c(1, 1, 2, 4)
+    )
+    expect_error(
+        complete_twins(rel_df),
+        "Multiple relationship codes in group 2"
+    )
+    expect_warning(
+        df <- complete_twins(rel_df, multi_code = "warn"),
+        "Multiple relationship codes in group 2"
+    )
+    expect_equal(dim(df), c(5, 5))
+    expect_equal(
+        as.character(df$code[df$id1 == "113" & df$id2 == "133"]),
+        "UZ twin"
+    )
+
+    expect_error(
+        complete_twins(rel_df, multi_code = "other"),
+        "Unknown multi_code argument"
+    )
+})
+